requirements.txt
aiogram
ccxt
pandas
pandas-ta
python-dotenv
apscheduler
